LayoutTests/js/dom/domjit-accessor-different-effect-expected.txt

Test DOMJIT accessors will handle different heap accesses correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test() is true
PASS successfullyParsed is true

TEST COMPLETE
